+++ src/main/java/com/ffbit/maven/solr/AbstractSolrMojo.java
@@ -1,10 +1,14 @@
 package com.ffbit.maven.solr;
 
-import com.ffbit.maven.solr.artefact.SolrArtifactResolver;
+import com.ffbit.maven.solr.artefact.ArtifactResolver;
+import com.ffbit.maven.solr.artefact.external.ExternalArtifacts;
+import com.ffbit.maven.solr.artefact.external.ExternalArtifactsFactory;
 import com.ffbit.maven.solr.extract.BootstrapExtractor;
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
+import org.apache.maven.plugins.annotations.Component;
+import org.apache.maven.plugins.annotations.Parameter;
 import org.sonatype.aether.RepositorySystem;
 import org.sonatype.aether.RepositorySystemSession;
 import org.sonatype.aether.artifact.Artifact;
@@ -17,47 +21,65 @@ import java.util.List;
 public abstract class AbstractSolrMojo extends AbstractMojo {
 
     /**
-     * @parameter property="contextPath" default-value="/"
+     * The context path for the server instance.
+     *
+     * @since 0.0.1
      */
+    @Parameter(property = "contextPath", defaultValue = "/")
     private String contextPath;
 
     /**
-     * @parameter property="port" default-value="8983"
+     * The running port.
+     *
+     * @since 0.0.1
      */
+    @Parameter(property = "port", defaultValue = "8983")
     private int port;
 
     /**
-     * @parameter property="solrVersion" default-value="4.3.0"
+     * The running Apache Solr version.
+     *
+     * @since 0.0.1
      */
+    @Parameter(property = "solrVersion", defaultValue = "4.3.0")
     private String solrVersion;
 
     /**
-     * @parameter property="solrHome" expression="${solr.solr.home}" default-value="${project.build.directory}/solr"
+     * The Apache Solr Home directory.
+     *
+     * @since 0.0.1
      */
+    @Parameter(property = "solrHome", alias = "${solr.solr.home}",
+            defaultValue = "${project.build.directory}/solr")
     private File solrHome;
 
     /**
      * The entry point to Aether, i.e. the component doing all the work.
      *
-     * @component
+     * @since 0.0.1
      */
+    @Component
     private RepositorySystem repositorySystem;
 
     /**
      * The current repository/network configuration of Maven.
      *
-     * @parameter default-value="${repositorySystemSession}"
-     * @readonly
+     * @since 0.0.1
      */
+    @Component
+    @Parameter(defaultValue = "${repositorySystemSession}")
     private RepositorySystemSession repositorySession;
 
     /**
      * The project's remote repositories to use for the resolution.
      *
-     * @parameter default-value="${project.remoteProjectRepositories}"
-     * @readonly
+     * @since 0.0.1
      */
+    @Parameter(defaultValue = "${project.remoteProjectRepositories}",
+            readonly = true)
     private List<RemoteRepository> remoteRepositories;
+
+    private ArtifactResolver artifactResolver;
 
     public String getContextPath() {
         return contextPath;
@@ -71,26 +93,43 @@ public abstract class AbstractSolrMojo extends AbstractMojo {
         return solrVersion;
     }
 
+    public File getSolrHome() {
+        return solrHome;
+    }
+
+    public void setRepositorySession(RepositorySystemSession repositorySession) {
+        this.repositorySession = repositorySession;
+    }
+
+    public void addRemoteRepository(RemoteRepository repository) {
+        remoteRepositories.add(repository);
+    }
+
     public String getArtifactPath() throws MojoExecutionException {
         return getArtifact().getFile().getAbsolutePath();
     }
 
-    public Artifact getArtifact() throws MojoExecutionException {
-        SolrArtifactResolver resolver =
-                new SolrArtifactResolver(repositorySystem, repositorySession, remoteRepositories);
-
-        return resolver.resolve(solrVersion);
+    private Artifact getArtifact() throws MojoExecutionException {
+        return artifactResolver.resolveSorlWarArtifact(solrVersion);
     }
 
     @Override
     public final void execute() throws MojoExecutionException, MojoFailureException {
-        BootstrapExtractor extractor = new BootstrapExtractor(solrHome, solrVersion);
+        artifactResolver = new ArtifactResolver(repositorySystem, repositorySession, remoteRepositories);
+        BootstrapExtractor extractor = new BootstrapExtractor(solrHome, solrVersion, artifactResolver);
         extractor.extract();
 
+        resolveExternalArtifacts();
         exportSolrHomeProperty();
         exportMavenLocalRepositoryProperty();
 
         executeGoal();
+    }
+
+    private void resolveExternalArtifacts() {
+        ExternalArtifactsFactory factory = new ExternalArtifactsFactory();
+        ExternalArtifacts externalArtifacts = factory.getExternalArtifacts(solrVersion);
+        artifactResolver.resolve(externalArtifacts.gerArtifacts());
     }
 
     private void exportMavenLocalRepositoryProperty() {